+++ examples/resources/iosxe_system/resource.tf
@@ -1,3 +1,4 @@
 resource "iosxe_system" "example" {
-  hostname = "ROUTER-1"
+  hostname             = "ROUTER-1"
+  ipv6_unicast_routing = true
 }