+++ internal/provider/resource_iosxe_restconf.go
@@ -149,9 +149,11 @@ func (r resourceRestconf) Read(ctx context.Context, req tfsdk.ReadResourceReques
 				// handle empty maps
 				if value.IsObject() && len(value.Map()) == 0 {
 					attributes[attr] = types.String{Value: ""}
-					continue
+				} else if value.Raw == "[null]" {
+					attributes[attr] = types.String{Value: ""}
+				} else {
+					attributes[attr] = types.String{Value: value.String()}
 				}
-				attributes[attr] = types.String{Value: value.String()}
 			}
 		}
 		state.Attributes.Elems = attributes